refactor(server): extract publishing dispatcher + pipedream client (3/3)
The publishing finale, and the last planned route group. Splits the
deferred publishing subsystem's hardest piece into 2 modules.

- src/server/routes/publishing-publish.js — generate-post-copy + the
  ~1,129-line publish dispatcher (per-channel fan-out: LinkedIn/X/Reddit/
  Facebook/Ghost/Medium/My Website) + runScheduledPublishes (the
  scheduled-publish runner, EXPORTED for the cron tick in server.js).
  Mounted `app.use('/api/publishing', publishingPublishRouter)` — both
  routes open (inline cron-bypass-OR-Clerk auth via jwtVerify).
- src/server/pipedream.js — the Pipedream Connect client (pipedreamProxy
  + getPipedreamAccessToken + getPipedreamAccountCredentials + the token
  cache). Its own module because pipedreamProxy is shared by the
  dispatcher AND the inline /api/admin/facebook/diag +
  /api/facebook/pipedream/list-pages routes (still in server.js). Exports
  pipedreamProxy; the two getPipedream* + cache vars stay internal.

server.js wiring: imports publishingPublishRouter (mount) +
runScheduledPublishes (cron call) + pipedreamProxy (the 2 inline FB
routes). Deps otherwise all already extracted (zernio/x/ghost/utm/text/
llm/db/auth + jose).

Process notes:
- A boundary over-reach swallowed the `_pd*` token-cache vars into the
  wrong module; lint (no-undef) caught it, moved them to pipedream.js.
- The route guard's parseImports doesn't handle the combined
  `import Default, { Named } from` form — it silently dropped the publish
  router's 2 routes (guard read 209). Split into two import lines (clean
  + idiomatic). KNOWN GUARD GAP worth a tiny follow-up to harden.

Test plan:
- node --check server.js + both modules -> OK
- npm run lint (no-undef) -> clean (caught _pd cache vars, generateHeroImage,
  createHmac, PORT)
- both modules boot-load (import()) -> resolve
- route-inventory guard -> 211 byte-identical (20 publishing paths: 14
  queue + 4 channels + 2 publish all reconstructed)
- vitest -> 93/93

Rollback: revert; restores inline dispatcher + helpers.
